appletb: Added hack to take over the touchbar device.
The hid driver core assigns devices to the hid-generic driver unless an
override is present in its hid_ignore_list. Unfortunately it does not
appear to be possible to override this behaviour dynamically. This
therefore introduces a terrible hack where we watch for iBridge device
and then forcefully unregister the current hid driver, allowing us to be
assigned as the driver when we (re)register ourselves as a hid driver.

To be removed when there is a better alternative...