appletb: Use a single structure to hold device info for both interfaces.
Logically the two USB interfaces presented are part of the same device,
the touchbar. And we treat them as part of one device. So change the
data structures such that we maintain a single structure per touchbar
device.

For now we assume a single device, but that assumption is only made at
the point of allocating and freeing the structure - everything else will
handle multiple devices.