appletb: Initial ambient light sensor support.
The iBridge also manages the ALS and exposes it as a standard HID light
sensor on the second USB HID interface. However, we cannot make use of
the existing driver support for such devices for several reasons:

* the hid-sensor-als driver is part of the hid-sensor-hub which in turn
  is a hid driver, and you can't have more than one hid driver per hid
  device.
* while the hid-sensors-als driver stores sensor readings received via
  interrupt in an iio buffer, reads on the sysfs
  .../iio:deviceX/in_illuminance_YYY attribute result in a get of the
  feature report; however, in the iBridge case the the illuminance field
  of that report is always 0. Instead, the input report needs to be
  requested.

For these and some other smaller reasons we (re)implement the necessary
report processing etc ourselves here. And because we do so, the
implementation is fairly bare bones. E.g. it does not implement iio
buffers and triggers, instead only supporting on-demand reads via the
sysfs in_illuminance_input attribute. But this is all that is needed for
desktops running the iio-sensor-proxy (used by Gnome and possibly
others).

This addresses Dunedan/mbp-2016-linux#16 .